requirements.txt
amqp==5.3.1
asgiref==3.8.1
astroid==3.3.6
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
dataclass-wizard==0.33.0
diff-match-patch==20241021
dill==0.3.9
distlib==0.3.9
Django==4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-debug-toolbar==4.4.6
django-filter==24.3
django-import-export==4.3.1
django-redis==5.4.0
django-rest-framework==0.1.0
django-simple-history==3.7.0
django-timezone-field==7.0
djangorestframework==3.15.2
drf-yasg==1.21.8
et_xmlfile==2.0.0
filelock==3.16.1
flake8==7.1.1
idna==3.10
inflection==0.5.1
isort==5.13.2
kombu==5.4.2
mailhog==0.1.1
mccabe==0.7.0
openpyxl==3.1.5
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
pycodestyle==2.12.1
pyflakes==3.2.0
pylint==3.3.2
python-crontab==4.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.2
tablib==3.7.0
tomlkit==0.13.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
virtualenv==20.28.0
wcwidth==0.2.13
